refactor: :truck: recording 관련 hook 폴더 생성
